+++ imports/ui/components/Activities.jsx
@@ -1,4 +1,4 @@
-import React, {Component } from 'react';
+import React, { Component } from 'react';
 import { MsgType } from './MsgType.jsx';
 import { Link } from 'react-router-dom';
 import Account from '../components/Account.jsx';
@@ -11,26 +11,36 @@ import _ from 'lodash';
 const T = i18n.createComponent();
 
 MultiSend = (props) => {
-    return <div>
-        <p><T>activities.senders</T>
-            <ul>
-                {props.msg.inputs.map((data,i) =>{
-                    return <li key={i}><Account address={data.address}/> <T>activities.sent</T> {data.coins.map((coin, j) =>{
-                        return <span key={j} className="text-success">{new Coin(coin.amount, coin.denom).toString(6)}</span>
+    return <Table striped className="mt-3">
+        <tbody>
+            <tr>
+                <th><T>common.from</T></th>
+                {props.msg.inputs.map((data, i) => {
+                    return <td key={i}><Account address={data.address} /></td>
+                })}
+            </tr>
+            <tr>
+                <th><T>common.to</T></th>
+                {props.msg.outputs.length > 1 ?
+                    <ol>{
+                        props.msg.outputs.map((data, i) => {
+                            return <li key={i} className="p-1"><Account address={data.address} />  <T>common.amount</T> {data.coins.map((coin, j) => {
+                                return <span key={j} className="text-success">{new Coin(coin.amount, coin.denom).toString(6)}</span>
+                            })}</li>
+                        })}
+                    </ol> : props.msg.outputs.map((data, i) => {
+                        return <td><Account address={data.address} /> {data.coins.map((coin, j) => {
+                            return <span key={j} className="font-weight-bold ml-4">{new Coin(coin.amount, coin.denom).toString(6)}</span>
+                        })}
+                        </td>
                     })}
-                    </li>
-                })}
-            </ul>
-            <T>activities.receivers</T>
-            <ul>
-                {props.msg.outputs.map((data,i) =>{
-                    return <li key={i}><Account address={data.address}/> <T>activities.received</T> {data.coins.map((coin,j) =>{
-                        return <span key={j} className="text-success">{new Coin(coin.amount, coin.denom).toString(6)}</span>
-                    })}</li>
-                })}
-            </ul>
-        </p>
-    </div>
+            </tr>
+            <tr>
+                <th><T>common.status</T></th>
+                <td>{!props.invalid ? <p className="text-danger">Failed</p> : <p className="text-success">Success</p>}</td>
+            </tr>
+        </tbody>
+    </Table>
 }
 
 export default class Activites extends Component {
@@ -65,15 +75,15 @@ export default class Activites extends Component {
             amount = msg.amount.map((coin) => new Coin(coin.amount, coin.denom).toString(6)).join(', ')
             return <p><Account address={msg.from_address} /> {(this.props.invalid)?<T>activities.failedTo</T>:''}<MsgType type={msg["@type"]} /> <span className="text-success">{amount}</span> <T>activities.to</T> <span className="address"><Account address={msg.to_address} /></span><T>common.fullStop</T></p>
         case "/cosmos.bank.v1beta1.MsgMultiSend":
-            return <div>
-                <Account address={msg.inputs[0].address} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+            return <p>
+                <Account address={msg.inputs[0].address} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> to {msg.outputs.length > 1 ? <T _purify={false} total={msg.outputs.length}>common.addressesCount</T> : <Account address={msg.outputs[0].address} /> }
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <MultiSend msg={msg} />
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
 
             // staking
         case "/cosmos.staking.v1beta1.MsgCreateValidator":
@@ -111,19 +121,19 @@ export default class Activites extends Component {
 
             // ibc
         case "/cosmos.IBCTransferMsg":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
         case "/cosmos.IBCReceiveMsg":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> 
-            </div>
+            </p>
 
             // IBC Client
         case "/ibc.core.client.v1.MsgCreateClient":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -136,14 +146,14 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.client.v1.MsgUpdateClient":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> 
-                <div className="d-inline">
-                    <span className="float-right"><i className="material-icons" onClick={this.toggle} >{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
-                    <Collapse isOpen={this.state.isOpen} style={{paddingBottom: this.state.isOpen ? "1rem" : "0"}}>
+                {this.props.showDetails ? <p className="d-inline"> 
+                    <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
+                    <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
                             <tbody>
                                 <tr>
@@ -157,14 +167,14 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.client.v1.MsgUpgradeClient": 
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
-                    <span className="float-right"><i className="material-icons" onClick={this.toggle} >{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
-                    <Collapse isOpen={this.state.isOpen} style={{ paddingBottom: this.state.isOpen ? "1rem" : "0" }}>
+                {this.props.showDetails ? <p className="d-inline">
+                    <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
+                    <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
                             <tbody>
                                 <tr>
@@ -182,14 +192,14 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.client.v1.MsgSubmitMisbehaviour":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
-                    <span className="float-right"><i className="material-icons" onClick={this.toggle} >{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
-                    <Collapse isOpen={this.state.isOpen} style={{ paddingBottom: this.state.isOpen ? "1rem" : "0" }}>
+                {this.props.showDetails ? <p className="d-inline">
+                    <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
+                    <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
                             <tbody>
                                 <tr>
@@ -203,8 +213,8 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.client.v1.Height":
             return <div>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
@@ -212,9 +222,9 @@ export default class Activites extends Component {
 
             // IBC Channel
         case "/ibc.core.channel.v1.MsgAcknowledgement":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -242,8 +252,8 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.Channel":
             return <div>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
@@ -257,9 +267,9 @@ export default class Activites extends Component {
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
             </div>
         case "/ibc.core.channel.v1.MsgChannelCloseConfirm":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -279,12 +289,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgChannelCloseInit":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -300,12 +310,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgChannelOpenAck":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -333,12 +343,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgChannelOpenConfirm":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -358,12 +368,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgChannelOpenInit":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -379,12 +389,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgChannelOpenTry":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -408,12 +418,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgRecvPacket":
-            return <div>
-                <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> 
-                <div className="d-inline">
+            return <p>
+                <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -433,22 +443,22 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.channel.v1.MsgTimeout":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
         case "/ibc.core.channel.v1.MsgTimeoutOnClose":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
 
             // IBC Connection
         case "/ibc.core.connection.v1.MsgConnectionOpenAck":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -472,12 +482,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.connection.v1.MsgConnectionOpenConfirm":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> 
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -494,12 +504,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>      
+                </p> : null}
+            </p>
         case "/ibc.core.connection.v1.MsgConnectionOpenInit":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -516,12 +526,12 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.connection.v1.MsgConnectionOpenTry":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} /> 
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -546,26 +556,26 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>    
-            </div>
+                </p> : null}
+            </p>
         case "/ibc.core.connection.v1.ConnectionEnd":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
         case "/ibc.core.connection.v1.Counterparty":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
         case "/ibc.core.connection.v1.Version":
-            return <div>
+            return <p>
                 <Account address={msg.signer} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type={msg["@type"]} />
-            </div>
+            </p>
 
-        // IBC Application
+            // IBC Application
         case "/ibc.applications.transfer.v1.MsgTransfer":
-            return <div>
+            return <p>
                 <Account address={msg.sender} /> {(this.props.invalid) ? <p className="ml-1"> <T>activities.failedTo</T></p> : ''}<MsgType type={msg["@type"]} /><T>activities.to</T> <span className="address"> <Account address={msg.receiver} /></span><T>common.fullStop</T>
-                <div className="d-inline">
+                {this.props.showDetails ? <p className="d-inline">
                     <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
                     <Collapse isOpen={this.state.isOpen}>
                         <Table striped className="mt-3">
@@ -581,11 +591,27 @@ export default class Activites extends Component {
                             </tbody>
                         </Table>
                     </Collapse>
-                </div>
-            </div>
+                </p> : null}
+            </p>
 
         default:
-            return <div><ReactJson src={msg} /></div>
+            return  <p>
+                <Account address={msg?.signer || msg?.sender || msg?.address || msg?.validator_address || msg?.delegator_address || msg?.proposer || msg?.depositor} /> {(this.props.invalid) ? <T>activities.failedTo</T> : ''}<MsgType type="Custom Message" />
+                {this.props.showDetails ? <p className="d-inline">
+                    <span className="float-right"><i className="material-icons" onClick={this.toggle}>{this.state.isOpen ? 'arrow_drop_down' : 'arrow_left'}</i></span>
+                    <Collapse isOpen={this.state.isOpen}>
+                        <Table className="mt-3">
+                            <tbody>
+                                <tr>
+                                    <th><T>common.message</T></th>
+                                    <td><ReactJson src={msg} /></td>
+                                </tr>
+                            </tbody>
+                        </Table>
+                    </Collapse>
+                </p> : null
+                }
+            </p>
         }
     }
 }